% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{get_db_kl_del}
\alias{get_db_kl_del}
\title{删除相似问列表}
\usage{
get_db_kl_del(conn_r = tsda::conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn_r}{连接}

\item{app_id}{程序}
}
\value{
返回值
}
\description{
删除相似问列表
}
\examples{
get_db_kl_del()
}
