% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{get_db_kn_del}
\alias{get_db_kn_del}
\title{删除相关数据}
\usage{
get_db_kn_del(conn_r = tsda::conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn_r}{连接}

\item{app_id}{程序}
}
\value{
返回值
}
\description{
删除相关数据
}
\examples{
get_db_kn_del()
}
