% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{get_kc_names}
\alias{get_kc_names}
\title{获取分类列表名称}
\usage{
get_kc_names(conn = tsda::conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}
}
\value{
返回值
}
\description{
获取分类列表名称
}
\examples{
get_kc_names()
}
