% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{get_kn_names}
\alias{get_kn_names}
\title{获取名称点的所有问题列表}
\usage{
get_kn_names(conn_r = tsda::conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn_r}{连接}

\item{app_id}{程序}
}
\value{
返回值
}
\description{
获取名称点的所有问题列表
}
\examples{
get_kn_names()
}
