% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_create}
\alias{kc_create}
\title{新建知识分类}
\usage{
kc_create(conn, app_id, kc_parentName, kc_name)
}
\arguments{
\item{conn}{连接信息}

\item{app_id}{程序代码}

\item{kc_parentName}{上级知识分类名称}

\item{kc_name}{本级知识分类名称}
}
\value{
返回值
}
\description{
新建知识分类
}
\examples{
kc_create()
}
