% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_delete}
\alias{kc_delete}
\title{删除知识分类}
\usage{
kc_delete(conn, app_id, kc_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{kc_name}{名称}
}
\value{
返回值
}
\description{
删除知识分类
}
\examples{
kc_delete()
}
