% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_getId}
\alias{kc_getId}
\title{获取知识分类id}
\usage{
kc_getId(conn, app_id, kc_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序名}

\item{kc_name}{知识分类名称}
}
\value{
返回值
}
\description{
获取知识分类id
}
\examples{
kc_getId()
}
