% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_getName}
\alias{kc_getName}
\title{获取知识分类名称}
\usage{
kc_getName(conn, app_id, kc_id)
}
\arguments{
\item{conn}{连接信息}

\item{app_id}{程序id}

\item{kc_id}{分类知识点id}
}
\value{
返回值
}
\description{
获取知识分类名称
}
\examples{
kc_getName()
}
