% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_getParentId}
\alias{kc_getParentId}
\title{获取上级知识分类的id}
\usage{
kc_getParentId(conn, app_id, kc_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序名称}

\item{kc_name}{知识分类名称}
}
\value{
返回值
}
\description{
获取上级知识分类的id
}
\examples{
kc_getParentId()
}
