% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_getParentName}
\alias{kc_getParentName}
\title{获取知识点的上级分类名称}
\usage{
kc_getParentName(conn, app_id, kc_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kc_name}{分类名称}
}
\value{
返回值
}
\description{
获取知识点的上级分类名称
}
\examples{
kc_getParentName()
}
