% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_query}
\alias{kc_query}
\title{知识库查询，直接从数据库中进行查询}
\usage{
kc_query(conn, app_id)
}
\arguments{
\item{conn}{连接信息}

\item{app_id}{程序id}
}
\value{
返回值
}
\description{
知识库查询，直接从数据库中进行查询
}
\examples{
kc_query()
}
