% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_update}
\alias{kc_update}
\title{知识分类类更新}
\usage{
kc_update(conn, app_id, old_kc_name, new_kc_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{old_kc_name}{旧分类名称}

\item{new_kc_name}{新分类名称}
}
\value{
返回值
}
\description{
知识分类类更新
}
\examples{
kc_update()
}
