% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kc.R
\name{kc_updateAll}
\alias{kc_updateAll}
\title{同步所有知识分类到数据库
后续根据情况进行更新}
\usage{
kc_updateAll(conn, app_id)
}
\arguments{
\item{conn}{数据库链接}

\item{app_id}{程序ID}
}
\value{
返回值
}
\description{
同步所有知识分类到数据库
后续根据情况进行更新
}
\examples{
kc_updateAll()
}
