% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{kk_getId}
\alias{kk_getId}
\title{标签答获取id}
\usage{
kk_getId(conn, app_id, kn_name, kk_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_name}{知识点名称}

\item{kk_name}{标签答名称}
}
\value{
返回值
}
\description{
标签答获取id
}
\examples{
kk_getId()
}
