% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{kk_getName}
\alias{kk_getName}
\title{获取标准答名称}
\usage{
kk_getName(conn, app_id, kk_id)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kk_id}{ID}
}
\value{
返回值
}
\description{
获取标准答名称
}
\examples{
kk_getName()
}
