% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{kk_push}
\alias{kk_push}
\title{提交标签答到数据库}
\usage{
kk_push(conn, app_id, kn_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_name}{知识点}
}
\value{
返回值
}
\description{
提交标签答到数据库
}
\examples{
kk_push()
}
