% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{kk_pushBatch}
\alias{kk_pushBatch}
\title{批量批取标准答}
\usage{
kk_pushBatch(
  conn_kms,
  conn_r = tsda::conn_rds("rdbe"),
  app_id = "caas",
  time = 0.02
)
}
\arguments{
\item{conn_kms}{连接py}

\item{conn_r}{连接r}

\item{app_id}{程序}

\item{time}{间隔时间默认1秒50个}
}
\value{
返回值
}
\description{
批量批取标准答
}
\examples{
kk_pushBatch
}
