% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R
\name{kk_update}
\alias{kk_update}
\title{更新标准答}
\usage{
kk_update(conn, app_id, kn_name, old_kk_name, new_kk_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_name}{知识点名称}

\item{old_kk_name}{原标准答}

\item{new_kk_name}{新标准答}
}
\value{
返回值
}
\description{
更新标准答
}
\examples{
kk_update()
}
