% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_create}
\alias{kl_create}
\title{创建相似问相关仙鹤}
\usage{
kl_create(conn, app_id, kn_name, kl_name = "test1212")
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序id}

\item{kn_name}{知识点名称}

\item{kl_name}{相似问名称}
}
\value{
返回值
}
\description{
创建相似问相关仙鹤
}
\examples{
kn_create()
}
