% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_delete}
\alias{kl_delete}
\title{删除知识叶}
\usage{
kl_delete(conn, app_id, kn_name, kl_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序id}

\item{kn_name}{知识点名称}

\item{kl_name}{知识页名称}
}
\value{
返回值
}
\description{
删除知识叶
}
\examples{
kl_delete()
}
