% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_getId}
\alias{kl_getId}
\title{根据名称获取相应的id}
\usage{
kl_getId(conn, app_id, kn_name, kl_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_name}{知识点名称}

\item{kl_name}{相似问名称}
}
\value{
返回值
}
\description{
根据名称获取相应的id
}
\examples{
kl_getId()
}
