% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_getName}
\alias{kl_getName}
\title{获取相似问问名称假设具有唯一性}
\usage{
kl_getName(conn, app_id, kl_id)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kl_id}{相似问ID}
}
\value{
返回值
}
\description{
获取相似问问名称假设具有唯一性
}
\examples{
kl_getName()
}
