% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kk.R, R/kl.R
\name{kl_push}
\alias{kl_push}
\title{提交标签答到数据库}
\usage{
kl_push(conn, app_id, kn_name, page = 1, page_size = 50, format = "list")

kl_push(conn, app_id, kn_name, page = 1, page_size = 50, format = "list")
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_name}{知识点名称}

\item{page}{页数}

\item{page_size}{每页大家}

\item{format}{格式}

\item{uag_id}{用户组默认0}
}
\value{
返回值

返回值
}
\description{
提交标签答到数据库

提交相似问更新到数据库
}
\examples{
kl_push()
kl_push()
}
