% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_pushBatch}
\alias{kl_pushBatch}
\title{针对相似问进行推送}
\usage{
kl_pushBatch(
  conn_kms,
  conn_r = tsda::conn_rds("rdbe"),
  app_id = "caas",
  time = 0.02
)
}
\arguments{
\item{conn_kms}{连接py}

\item{conn_r}{连接r}

\item{app_id}{程序}

\item{time}{时间}
}
\value{
返回值
}
\description{
针对相似问进行推送
}
\examples{
kl_pushBatch()
}
