% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_pushBatch_auto}
\alias{kl_pushBatch_auto}
\title{同步处理相似问的同步}
\usage{
kl_pushBatch_auto(
  conn_kms,
  conn_r = tsda::conn_rds("rdbe"),
  app_id = "caas",
  time = 0.02
)
}
\arguments{
\item{conn_kms}{py连接}

\item{conn_r}{r连接}

\item{app_id}{程序}

\item{time}{时间}
}
\value{
返回值
}
\description{
同步处理相似问的同步
}
\examples{
kl_pushBatch_auto()
}
