% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_update}
\alias{kl_update}
\title{更新相似问}
\usage{
kl_update(conn, app_id, kn_name, old_kl_name, new_kl_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_name}{知识点名称}

\item{old_kl_name}{原相似问名称}

\item{new_kl_name}{新相似问名称}
}
\value{
返回值
}
\description{
更新相似问
}
\examples{
kl_update()
}
