% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{kn_delete}
\alias{kn_delete}
\title{删除知识点}
\usage{
kn_delete(conn, app_id, kn_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序id}

\item{kn_name}{知识点名称}
}
\value{
返回值
}
\description{
删除知识点
}
\examples{
kn_delete()
}
