% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{kn_getId}
\alias{kn_getId}
\title{知识点获取相应id}
\usage{
kn_getId(conn, app_id, kn_name)
}
\arguments{
\item{conn}{数据库}

\item{app_id}{程序}

\item{kn_name}{知识点名称}
}
\value{
返回知识点ID
}
\description{
知识点获取相应id
}
\examples{
kn_getId()
}
