% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{kn_getName}
\alias{kn_getName}
\title{知识点获取名称}
\usage{
kn_getName(conn, app_id, kn_id)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_id}{知识点ID}
}
\value{
返回知识点名称
}
\description{
知识点获取名称
}
\examples{
kn_getName()
}
