% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{kn_push}
\alias{kn_push}
\title{提交知识点更新到数据库}
\usage{
kn_push(conn, app_id, kc_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序id}

\item{kc_name}{知识分类名称}
}
\value{
返回值
}
\description{
提交知识点更新到数据库
}
\examples{
kn_push()
}
