% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{kn_pushBatch}
\alias{kn_pushBatch}
\title{批量批送相关信息}
\usage{
kn_pushBatch(conn_kms, conn_rdb = tsda::conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn_kms}{通过py连接数据库}

\item{conn_rdb}{通过r连接数据库}

\item{app_id}{程序名称}
}
\value{
返回值
}
\description{
批量批送相关信息
}
\examples{
kn_pushBatch()
}
