% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{kn_pushBatch_auto}
\alias{kn_pushBatch_auto}
\title{批量更新知识点用于自动同}
\usage{
kn_pushBatch_auto(conn_kms, conn_rdb = tsda::conn_rds("rdbe"), app_id = "caas")
}
\arguments{
\item{conn_kms}{py连接}

\item{conn_rdb}{r连接}

\item{app_id}{程序}
}
\value{
返回值
}
\description{
批量更新知识点用于自动同
}
\examples{
kn_pushBatch_auto()
}
