% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kn.R
\name{kn_update}
\alias{kn_update}
\title{更新知识点}
\usage{
kn_update(conn, app_id, old_kn_name, new_kn_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序id}

\item{old_kn_name}{原知识点名称}

\item{new_kn_name}{新知识点名称}
}
\value{
返回值
}
\description{
更新知识点
}
\examples{
kn_update()
}
