% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl_push}
\alias{kl_push}
\title{提交相似问更新到数据库}
\usage{
kl_push(conn, app_id, kn_name)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序}

\item{kn_name}{知识点名称}
}
\value{
返回值
}
\description{
提交相似问更新到数据库
}
\examples{
kl_push()
}
